+++ build_test.sh
@@ -4,21 +4,18 @@ export PATH=\
 /opt/buildout.python/bin:\
 $PATH:
 
-if [[ "$1" = "plone-4.3" ]]
-then
-    config=buildout-plone-4.3.cfg
-fi
+echo $WEBDAV_URL
+echo $PLONE_VERSION
+echo $DOCKER
 
-if [[ "$1" = "plone-5.0" ]]
-then
-    config=buildout-plone-5.0.cfg
-fi
+config=buildout-plone-$PLONE_VERSION.cfg
 
-export WEBDAV_URL=http://demo.xml-director.info:22081/exist/webdav/db
+docker pull $DOCKER
+docker run -d -p 127.0.0.1:8080:8080 $DOCKER
 
-virtualenv-2.7 .
-bin/pip install -U setuptools==7.0  
-bin/python bootstrap.py -c $config --setuptools-version 7.0 --version 2.2.5
+#virtualenv-2.7 .
+pip install -U setuptools==7.0  
+python bootstrap.py -c $config --setuptools-version 18.4 --version 2.2.5
 bin/buildout -c $config
 bin/test -s xmldirector.plonecore
 ##bin/coverage run bin/test