+++ xmldirector/plonecore/__init__.py
@@ -19,3 +19,5 @@ for mod, min_version in [('lxml', [3, 4])]:
         min_version_s = '.'.join(map(str, min_version))
         raise RuntimeError('xmldirector.plonecore requires module {} in version {} or higher (installed: {})'.format(
             mod, min_version_s, dist.version))
+
+