+++ xmldirector/plonecore/connector.py
@@ -19,17 +19,16 @@ from plone.supermodel import model
 from plone.registry.interfaces import IRegistry
 
 from xmldirector.plonecore.i18n import MessageFactory as _
-from xmldirector.plonecore.davfs import DAVFSWrapper as DAVFS
 from xmldirector.plonecore.interfaces import IWebdavSettings
-
+from xmldirector.plonecore.fswrapper import get_fs_wrapper
 from xmldirector.plonecore.logger import LOG
 
 
 class IConnector(model.Schema):
 
-    webdav_subpath = schema.TextLine(
-        title=_(u'Subdirectory in Exist-DB'),
-        description=_(u'Subdirectory in Exist-DB'),
+    webdav_url = schema.TextLine(
+        title=_(u'(optional) connection URL of storage'),
+        description=_(u'WebDAV: http://host:port/path/to/webdav, Local filesystem: file://path/to/directory, AWS S3: s3://bucketname, SFTP sftp://host/path, FTP: ftp://host/path'),
         required=False
     )
 
@@ -40,6 +39,12 @@ class IConnector(model.Schema):
 
     webdav_password = schema.TextLine(
         title=_(u'(optional) password overriding the system settings'),
+        required=False
+    )
+
+    webdav_subpath = schema.TextLine(
+        title=_(u'Subdirectory relative to the global connection URL'),
+        description=_(u'Use this value for configuring a more specific subpath'),
         required=False
     )
 
@@ -102,12 +107,19 @@ class Connector(Item):
             password = adapted.webdav_password or ''
 
         try:
-            return DAVFS(url, credentials=dict(username=username,
-                                               password=password))
+            return get_fs_wrapper(url, credentials=dict(username=username, password=password))
         except fs.errors.ResourceNotFoundError:
-            LOG.error(u'Error accessing {}::{}::{}'.format(self.absolute_url(), url, self.REQUEST.get('HTTP_USER_AGENT')), exc_info=True)
+            LOG.error(u'Error accessing {}::{}::{}'.format(
+                self.absolute_url(), url, self.REQUEST.get('HTTP_USER_AGENT')), exc_info=True)
             raise zExceptions.Unauthorized(url)
+        except fs.errors.ResourceInvalidError:
+            parts = url.rsplit('/', 1)
+            wrapper = get_fs_wrapper(parts[0], credentials=dict(username=username, password=password))
+            wrapper.__leaf__ = True
+            wrapper.__leaf_filename__ = parts[1]
+            return wrapper
         except Exception as e:
-            LOG.error(u'Error accessing {}::{}::{}'.format(self.absolute_url(), url, self.REQUEST.get('HTTP_USER_AGENT')), exc_info=True)
+            LOG.error(u'Error accessing {}::{}::{}'.format(
+                self.absolute_url(), url, self.REQUEST.get('HTTP_USER_AGENT')), exc_info=True)
             e.url = url
             raise e